Removed an unused authentication method in UserFactory checkpassword(String email, String password).
Refactored all user related classes to use the "passwordHash" attribute instead of "password" to make it clear we are only dealing with hashes.

Updated the UpdateUserInformation page to not auto-fill with the password, and if nothing is entered into the password field by the user their password is kept the same. Updated the view to reflect this.